// types/task.interface.ts
export interface ITask {
    id: string; // order id
    title: string; // Khách Hang ABC đã mua 5 sản phẩm
    label: string; // hình thức thanh toán
    status: string; // order status
    employee?: string;
    total_price: string;
    priority: string; // đơn vị giao hàng
    create: string;
}